requirements.txt
torch>=1.10.0
PyYAML==6.0
datasets==2.14.5
tokenizers==0.13.3
matplotlib==3.7.2
tqdm==4.66.1
pyarrow==14.0.1
numpy==1.26.4
